backend/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-environ
gunicorn
psycopg2-binary
pandas
openpyxl
whitenoise
django-cors-headers
Pillow
python-docx
PyPDF2
python-dotenv